+++ docs/fr_FR/changelog.md
@@ -2,9 +2,19 @@
 
 *[Documentation](index.md)*
 
+### 2018-05-02
+  - Ajout d'une option pour le format de date dans le log (format php).
+  - Import infos: bugfix #cmd#
+  - Import infos: Ajout d'un champ Rechercher.
+  - Import infos: Ajout d'un texte commande historisée ou non.
+  - Ajout de log en debug.
+  - Mise à jour de la doc.
+
 ### 2018-04-30
 
 - Ajout d'une option d'importation de commandes infos.
+- Bugfix affichage de la checkbox Inverser sous Chrome.
+- Nouvelle option pour afficher ou non la date/heure de mise à jour du log.
 
 ### 2018-04-29